Allow different runner implementation per platform
Windows ohai are very different than Linux one, so we need different implementation
Using modules is quite efficient, it allows code seperation and override.
